requirements.txt
langchain-text-splitters
# Core LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-ollama>=1.0.0
langchain-huggingface>=1.0.0

# Vector store
faiss-cpu>=1.7.4

# Embeddings (free local models)
sentence-transformers>=5.0.0

# PDF processing
pypdf>=3.17.0
pdfplumber>=0.10.0  # Better table extraction for accounting documents

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# OpenAI API
openai>=1.12.0

# Optional: Streamlit for UI
streamlit>=1.31.0

# Optional: Evaluation
ragas>=0.1.0

# Utility
python-dotenv>=1.0.0
pydantic>=2.0.0
